Expand mark_as_visited function to also return new level status

+++ app/attractions/views.py
@@ -134,37 +134,43 @@ def get_attraction(attraction_id):
         "attraction.html", attraction=attraction, visited=visited, groups=groups
     )
 
-
 # Function for marking an attraction as visited
 @attractions.route("/attractions/<int:attraction_id>/mark_as_visited", methods=["POST"])
 @login_required
 def mark_as_visited(attraction_id):
     data = request.get_json()
     visited = data.get("visited", False)
-
-    if visited:
-        # Mark attraction as visited
-        existing_record = VisitedAttraction.query.filter_by(
-            user_id=current_user.id, attraction_id=attraction_id
+    existing_record = VisitedAttraction.query.filter_by(
+        user_id=current_user.id, attraction_id=attraction_id
         ).first()
-        if not existing_record:
-            visited_attraction = VisitedAttraction(
-                user_id=current_user.id, attraction_id=attraction_id
-            )
-            db.session.add(visited_attraction)
-            db.session.commit()
-            return jsonify(
-                {"status": "success", "message": "Attraction marked as visited."}
-            )
-
-    else:
-        existing_record = VisitedAttraction.query.filter_by(
-            user_id=current_user.id, attraction_id=attraction_id
-        ).first()
-        if existing_record:
-            db.session.delete(existing_record)
-            db.session.commit()
-            return jsonify({"status": "success", "message": "Attraction mark removed."})
+    
+    # Mark attraction as visited
+    if visited and not existing_record:
+        visited_attraction = VisitedAttraction(
+            user_id=current_user.id, attraction_id=attraction_id)
+        db.session.add(visited_attraction)
+        db.session.commit()
+        level = current_user.level
+        return jsonify({"status": "success", 
+                        "message": "Attraction mark removed.",
+                        "current_level": level['current_level'],
+                        "points_required": level['points_required'],
+                        "points_missing": level['points_missing'],
+                        "progress": level['progress']
+                        })
+    
+    # Unmark attraction as visited
+    elif not visited and existing_record:
+        db.session.delete(existing_record)
+        db.session.commit()
+        level = current_user.level
+        return jsonify({"status": "success", 
+                        "message": "Attraction mark removed.",
+                        "current_level": level['current_level'],
+                        "points_required": level['points_required'],
+                        "points_missing": level['points_missing'],
+                        "progress": level['progress']
+                        })
 
     return jsonify({"status": "error", "message": "An error occurred."})
